requirements.txt
streamlit==1.20.0
langchain==0.0.1
langchain_groq==0.0.1
langchain_google_genai==0.0.1
langchain_community==0.0.1
faiss-cpu==1.7.3
python-dotenv==1.0.0
